First rough version of the fixer
Add the initial version of the helper along with a little guide on how to use it
and a Makefile for the uninitiated.

The helper currently only accepts internal tlds from the commandline, the PHP
implemenation that was proposed in the workaround in this chromium bug
https://bugs.chromium.org/p/chromium/issues/detail?id=30636 does accept the
TLDs list as a query parameter, we may add this in the future.